refactor(pipeline): move some ident logic in identifier
